+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -201,7 +201,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-04-15T13:08:37.542Z",
+    "2025-04-15T14:53:32.746Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -241,7 +241,7 @@ export const RetrieveDatasetResponseBody$outboundSchema: z.ZodType<
   updatedById: z.string().optional(),
   metadata: z.lazy(() => RetrieveDatasetMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-04-15T13:08:37.542Z"))
+  updated: z.date().default(() => new Date("2025-04-15T14:53:32.746Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {